Move NodeNG to its own file
NodeNG is getting really big. Time to refactor node_classes.py
